refactor(SDK-1120): nest onboarding Contractor Address under onboarding/
Aligns Contractor/Address with the Employee domain's onboarding/management/shared
convention, where the older onboarding component previously sat at the domain root
instead of alongside its new management/shared siblings. Pure move; no behavior change.
