--- src/components/Contractor/Address/index.ts
@@ -1,2 +1,0 @@
-export { Address, type AddressProps } from './Address'
-export type { AddressDefaultValues } from './types'
