+++ src/components/Contractor/Address/onboarding/Address.tsx
@@ -2,9 +2,9 @@ import type { JSX } from 'react'
 import { useTranslation } from 'react-i18next'
 import classNames from 'classnames'
 import type { FallbackProps } from 'react-error-boundary'
+import { useContractorAddressForm } from '../shared/useContractorAddressForm'
+import type { ContractorAddressOptionalFieldsToRequire } from '../shared/useContractorAddressForm'
 import styles from './Address.module.scss'
-import { useContractorAddressForm } from './shared/useContractorAddressForm'
-import type { ContractorAddressOptionalFieldsToRequire } from './shared/useContractorAddressForm'
 import type { AddressDefaultValues } from './types'
 import { BaseBoundaries, BaseLayout } from '@/components/Base'
 import { SDKFormProvider } from '@/partner-hook-utils/form/SDKFormProvider'
