+++ src/components/Contractor/OnboardingFlow/OnboardingFlowComponents.tsx
@@ -1,12 +1,12 @@
 import { useTranslation } from 'react-i18next'
 import { ContractorList } from '../ContractorList/onboarding/ContractorList'
 import { ContractorProfile } from '../Profile/onboarding/ContractorProfile'
-import { Address } from '../Address'
+import { Address } from '../Address/onboarding/Address'
 import { PaymentMethod } from '../PaymentMethod/PaymentMethod'
 import { NewHireReport } from '../NewHireReport/NewHireReport'
 import { ContractorSubmit } from '../Submit/Submit'
 import type { ContractorDetailsFormData } from '../Profile/shared'
-import type { AddressDefaultValues } from '../Address'
+import type { AddressDefaultValues } from '../Address/onboarding/types'
 import { useFlow, type FlowContextInterface } from '@/components/Flow/useFlow'
 import type { RequireAtLeastOne } from '@/types/Helpers'
 import type { BaseComponentInterface } from '@/components/Base'
