+++ src/components/Contractor/SelfOnboardingFlow/SelfOnboardingComponents.tsx
@@ -4,7 +4,7 @@ import { useFlow } from '@/components/Flow/useFlow'
 import type { BaseComponentInterface } from '@/components/Base'
 import { Landing as LandingComponent } from '@/components/Contractor/Landing'
 import { ContractorProfile as ProfileComponent } from '@/components/Contractor/Profile/onboarding/ContractorProfile'
-import { Address as AddressComponent } from '@/components/Contractor/Address'
+import { Address as AddressComponent } from '@/components/Contractor/Address/onboarding/Address'
 import { PaymentMethod as PaymentMethodComponent } from '@/components/Contractor/PaymentMethod/PaymentMethod'
 import { DocumentSigner as DocumentSignerComponent } from '@/components/Contractor/Documents/DocumentSigner'
 import { OnboardingSummary as OnboardingSummaryComponent } from '@/components/Contractor/OnboardingSummary'
