+++ src/components/Contractor/exports/contractorOnboarding.ts
@@ -20,8 +20,8 @@ export {
   type ContractorProfileSelfOnboardingProps,
 } from '../Profile/onboarding/ContractorProfile'
 export type { ContractorDetailsFormData as ContractorProfileFormData } from '../Profile/shared'
-export { Address, type AddressProps } from '../Address'
-export type { AddressDefaultValues } from '../Address/types'
+export { Address, type AddressProps } from '../Address/onboarding/Address'
+export type { AddressDefaultValues } from '../Address/onboarding/types'
 export { PaymentMethod } from '../PaymentMethod/PaymentMethod'
 export type { PaymentMethodProps } from '../PaymentMethod/types'
 export { NewHireReport } from '../NewHireReport/NewHireReport'
